[cherry-pick] move norm from arm to host; fix fill_any_like (#5968) (#5975)
test=develop
