+++ lite/backends/arm/math/norm.cc
@@ -23,29 +23,6 @@ namespace lite {
 namespace arm {
 namespace math {
 
-void norm(const float* input,
-          const int pre_n,
-          const int n,
-          const int post_n,
-          const float epsilon,
-          float* out,
-          Context<TARGET(kARM)>* ctx) {
-  for (int i = 0; i < pre_n; i++) {
-    for (int k = 0; k < post_n; k++) {
-      float sum = epsilon;
-      const float* in_tmp = input + i * n * post_n + k;
-      for (int j = 0; j < n; j++) {
-        sum += in_tmp[j * post_n] * in_tmp[j * post_n];
-      }
-      sum = std::sqrt(sum);
-      float* out_tmp = out + i * n * post_n + k;
-      for (int j = 0; j < n; j++) {
-        out_tmp[j * post_n] = in_tmp[j * post_n] / sum;
-      }
-    }
-  }
-}
-
 void matrix_norm_row(const float* x_data,
                      const float* scale_data,
                      const float* bias_data,
@@ -181,41 +158,6 @@ void matrix_norm_row(const float* x_data,
       ++x_ptr;
     }
   }  // for bi
-}
-
-void p_norm(const float* input,
-            const int pre_n,
-            const int n,
-            const int post_n,
-            const float epsilon,
-            float* out,
-            const int porder) {
-  // porder == 0 : count the non-zero elements of inputs
-  if (porder == 0) {
-    for (int i = 0; i < pre_n; i++) {
-      for (int k = 0; k < post_n; k++) {
-        float sum = epsilon;
-        const float* in_tmp = input + i * n * post_n + k;
-        for (int j = 0; j < n; j++) {
-          sum += in_tmp[j * post_n] != 0;
-        }
-        float* out_tmp = out + i * post_n + k;
-        *out_tmp = sum;
-      }
-    }
-  } else {
-    for (int i = 0; i < pre_n; i++) {
-      for (int k = 0; k < post_n; k++) {
-        float sum = epsilon;
-        const float* in_tmp = input + i * n * post_n + k;
-        for (int j = 0; j < n; j++) {
-          sum += std::pow(in_tmp[j * post_n], porder);
-        }
-        float* out_tmp = out + i * post_n + k;
-        *out_tmp = std::pow(sum, 1.f / porder);
-      }
-    }
-  }
 }
 
 }  // namespace math
