+++ lite/backends/host/math/CMakeLists.txt
@@ -4,6 +4,7 @@ lite_cc_library(math_host SRCS
     slice.cc
     split.cc
     gpc.cc
+    norm.cc
     pad3d.cc
     concat.cc
     stack.cc
