+++ lite/kernels/host/CMakeLists.txt
@@ -23,6 +23,7 @@ add_kernel(assign_value_compute_host Host basic SRCS assign_value_compute.cc DEP
 # extra kernels
 add_kernel(reduce_all_compute_host Host extra SRCS reduce_all_compute.cc DEPS ${lite_kernel_deps} math_host)
 add_kernel(unstack_compute_host Host extra SRCS unstack_compute.cc DEPS ${lite_kernel_deps})
+add_kernel(norm_compute_host Host extra SRCS norm_compute.cc DEPS ${lite_kernel_deps} math_host)
 add_kernel(shape_compute_host Host extra SRCS shape_compute.cc DEPS ${lite_kernel_deps})
 add_kernel(is_empty_compute_host Host extra SRCS is_empty_compute.cc DEPS ${lite_kernel_deps})
 add_kernel(crf_decoding_compute_host Host extra SRCS crf_decoding_compute.cc DEPS ${lite_kernel_deps})
